update placeholder text and add test
- moved tests to test_process (since it is related)
- use better placeholder text, and add missing placeholders to configs
